rock-paper-scissors/hexs2pngs/requirements.txt
# Empty
pillow



